CFD: Create cases from user-editable scheme templates
The idea is to have fvSolution and fvSchemes data in a best-practice
database that is available to users to edit. At present this
just takes the form of copying fvSchemes and fvSolution from
this location, but in future different dictionaries could be
merged in order to easily maintain best-practice data for different
functionalities.

The data is stored in <Cfd_mod_dir>/data/defaults under the solver name.